refactor: modular conversational agents
